refactor: move InternalClockTester
